requirements.txt
Pillow
imagehash
hwp-extract
pytest
olefile
scikit-image
opencv-python
numpy
pywin32